refactor: adjust imports for error type references
