+++ GitHubClients/TagClient.ts
@@ -3,7 +3,7 @@ import { TagModel } from "../deps.ts";
 import { Utils } from "../deps.ts";
 import { GitHubHttpStatusCodes } from "../core/Enums.ts";
 import { GitHubClient } from "../deps.ts";
-import { TagError } from "./Errors/TagError.ts";
+import { TagError } from "../deps.ts";
 
 /**
  * Provides a client for interacting with GitHub GIT tags.
